chore: code refactor
