+++ web/core/components/issues/attachment/attachment-list-item.tsx
@@ -90,7 +90,7 @@ export const IssueAttachmentsListItem: FC<TIssueAttachmentsListItem> = observer(
               </>
             )}
 
-            <CustomMenu ellipsis closeOnSelect placement="bottom-end" openOnHover disabled={disabled}>
+            <CustomMenu ellipsis closeOnSelect placement="bottom-end" disabled={disabled}>
               <CustomMenu.MenuItem
                 onClick={(e) => {
                   e.preventDefault();
